legacy_python/requirements.txt
# 星影 - Hoshikage
# 依存関係リスト（CUDA対応版）

# Webフレームワーク
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# 環境変数管理
python-dotenv>=1.0.0

# LLM推論エンジン（CUDA対応版）
# llama-cpp-python>=0.2.0; sys_platform != 'win32'

# ベクトルデータベース
chromadb>=0.4.0

# 埋め込みモデル
sentence-transformers>=2.2.0
sentencepiece>=0.1.99

# 日本語処理
fugashi>=1.3.0
unidic-lite>=1.0.8

# 機械学習
scikit-learn>=1.3.0
numpy>=1.24.0

# ユーティリティ
requests>=2.31.0
